Add initial implementation of the library.
The whole thing is pretty simple, you instantiate a BinaryDownloader
which contains the state of the thing and then you pass an argument
parser through it and eventually call BinaryDownloader.execute to
run it.

I need to write something that uses this to figure out if the API
is exactly what I want or not.