Remove conflicting $XDG_CACHE_HOME/{binary_name} if applicable.
This is a complex issue. This library is an outgrowth of a pattern
that became common to several of my pre-commit hooks around downloading
binaries. Some of the early versions of those hooks would save the
binary without the version string as part of the path, commonly
at $XDG_CACHE_HOME/{binary_name}. Some older systems will still have
this binary and it will conflict with the new paths at
$XDG_CACHE_HOME/{binary_name}/{version}. Rather than just fail
and force a user to clean up the mess we instead proactively clean
up ourselves.

This means those older pre-commit hooks will fail if they are run
after newer hooks using this library. That's a sacrifice they're
willing to make.