Use OS-level locks rather than timing for parallelism conflicts
My previous major change was to introduce a retry when downloading
the binary in the hopes that it would allow multiple instances of
any clients of this library to run in parallel without coordination.

It did, a bit, but it was a quick'n'dirty way of accomplishing it
and didn't solve the problem.

Now instead we use OS-level locks on the file which is much safer
and thorough and correct. This also makes the code Linux-only.